+++ android/src/main/java/com/bitmovin/player/reactnative/RNPlayerView.kt
@@ -16,6 +16,7 @@ import com.bitmovin.player.api.event.SourceEvent
 import com.bitmovin.player.api.ui.PlayerViewConfig
 import com.bitmovin.player.api.ui.StyleConfig
 import com.bitmovin.player.reactnative.converter.toJson
+import com.bitmovin.player.reactnative.extensions.playerModule
 import com.facebook.react.ReactActivity
 import com.facebook.react.bridge.*
 import com.facebook.react.uimanager.events.RCTEventEmitter
@@ -102,8 +103,23 @@ class RNPlayerView(
     private val activityLifecycle = (context.currentActivity as? ReactActivity)?.lifecycle
         ?: error("Trying to create an instance of ${this::class.simpleName} while not attached to a ReactActivity")
 
+    /**
+     * Relays the provided set of events, emitted by the player, together with the associated name
+     * to the `eventOutput` callback.
+     */
+    private var playerEventRelay: EventRelay<Player, Event> = EventRelay<Player, Event>(
+        EVENT_CLASS_TO_REACT_NATIVE_NAME_MAPPING,
+        ::emitEventFromPlayer,
+    )
+
+    internal var enableBackgroundPlayback: Boolean = false
+    var playerInMediaSessionService: Player? = null
+
     private val activityLifecycleObserver = object : DefaultLifecycleObserver {
         override fun onStart(owner: LifecycleOwner) {
+            if (playerInMediaSessionService != null) {
+                playerView?.player = playerInMediaSessionService
+            }
             playerView?.onStart()
         }
 
@@ -116,10 +132,28 @@ class RNPlayerView(
         }
 
         override fun onStop(owner: LifecycleOwner) {
+            removePlayerForBackgroundPlayback()
             playerView?.onStop()
         }
 
         override fun onDestroy(owner: LifecycleOwner) = dispose()
+
+        // When background playback is enabled,
+        // remove player from view so it does not get paused when entering background
+        private fun removePlayerForBackgroundPlayback() {
+            playerInMediaSessionService = null
+            val player = playerView?.player ?: return
+
+            if (!enableBackgroundPlayback) {
+                return
+            }
+            if (context.playerModule?.mediaSessionPlaybackManager?.player != player) {
+                return
+            }
+
+            playerInMediaSessionService = player
+            playerView?.player = null
+        }
     }
 
     init {
@@ -132,15 +166,6 @@ class RNPlayerView(
 
         activityLifecycle.addObserver(activityLifecycleObserver)
     }
-
-    /**
-     * Relays the provided set of events, emitted by the player, together with the associated name
-     * to the `eventOutput` callback.
-     */
-    private val playerEventRelay = EventRelay<Player, Event>(
-        EVENT_CLASS_TO_REACT_NATIVE_NAME_MAPPING,
-        ::emitEventFromPlayer,
-    )
 
     /**
      * Relays the provided set of events, emitted by the player view, together with the associated name